// react-vite-jobhunter-chuyendeweb/src/types/backend.d.ts
interface IBackendRes<T> {
    status: number;
    message: string;
    data: T;
    error?: string | string[]; // Thêm trường error, có thể là string hoặc mảng string
}